ok right now I got the bullets to shoot but really really badly. There are visually 2 bullets traveling at the same time even though ive only clicked once. However it somewhat resembles the prototype now it's just the enemy regestering being hit as the thing i haven't implemented. I'll also try to fix the problem I described tomorrow but it seems im done for the day

+++ My NEA Project/Program.cs
@@ -8,8 +8,8 @@ namespace My_NEA_Project
 {
     public struct Movement
     {
-        public int horrizontalMovement;
-        public int verticalMovement;
+        public double horrizontalMovement;
+        public double verticalMovement;
     }
 
     public struct Material
